ai-todo-app/pyproject.toml
[project]
name = "ai-todo-app"
version = "0.1.0"
description = "A structured layered CRUD Todo API built with FastAPI and SQLModel"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.6.0",
    "uvicorn[standard]>=0.32.0",
    "psycopg[binary]>=3.2.0",
    "openai-agents[sqlalchemy]>=0.0.7",
    "fastmcp>=3.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
